Move binning scripts into one script
